VCV-213: remove duplicated review state file
